Migrate default textures out of ib_Core and into the rendergraph API

+++ Iceberg/Source/iceberg/ib_util.c
@@ -50,7 +50,7 @@ uint32_t ib_firstBitLowU32(uint32_t value)
 
 uint32_t ib_bitCountU64(uint32_t value)
 {
-	return _mm_popcnt_u64(value);
+	return (uint32_t)_mm_popcnt_u64(value);
 }
 
 
@@ -70,5 +70,5 @@ uint32_t ib_firstBitLowU64(uint64_t value)
 
 uint32_t ib_bitCountU32(uint64_t value)
 {
-	return _mm_popcnt_u64(value);
+	return (uint32_t)_mm_popcnt_u64(value);
 }
